Finished the project
I finished the project with the logic and updated the GUI to make it so it is much easier to manage than the lines I drew on a canvas in my last iteration. The GUI still isn't super good looking and has some bugs. The algorithm mixed with the GUI also leads to very slow speeds and the program eventually crashing so I am going to include the same code but without a GUI to show that it does work.